// src/server/app/framework/search-engine/models/input-model/applied-filter.ts
import { ESort } from './sort-enum';
import { EList } from './list-enum';
export class AppliedFilter {
  maxSalary: number;
  minSalary: number;
  minExperience: number;
  maxExperience: number;
  proficiencies: string[] = new Array(0);
  interestedIndustries: string[]= new Array();
  education: string[] = new Array(0);
  //industry: string[] = new Array(0);
  joinTime: string;
  location: string;
  companySize: string;
  mustHaveComplexity: boolean = false;
  sortBy: ESort = ESort.BEST_MATCH;
  listName: EList;
}